% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_import.R
\name{bulk_import}
\alias{bulk_import}
\title{Bulk import SPSS(sav) and STATA(dta) and MS Excel(csv) data into R}
\usage{
bulk_import(data_dir = NULL, file_type = "all")
}
\arguments{
\item{data_dir}{The folder storing data files.}

\item{file_type}{The data file type. Options could be 'sav' or 'dta' or 'csv'. Default is 'all'.
\itemize{
\item sav uses \code{read_sav()} from 'haven' package.
\item dta uses \code{read_dta()} from 'haven' package.
\item csv uses \code{read.csv()} from 'utils' package.
\item The default 'all' imports all three kinds of data.
}}
}
\value{
Two results as follows:
\itemize{
\item Import SPSS files in \code{data_dir} to R global environment;
\item List names of tables imported to R.
}
}
\description{
Many NEAR databases stroe data in SPSS or STATA and MS Excel. Therefore chances are that we need
to import many different files at a time in R for further analysis.
This function is to bulk import data files of types mentioned above into R environment.
}
\note{
\itemize{
\item \code{here} function is a fast and neat way of defining data directory \code{data_dir}.
\item Be careful when importing the csv files. It could be messy since this function just uses default parameters of \code{read.csv()}.
}
}
\examples{
\dontrun{
# Assume the data files are in path:'data/raw/SNAC-K'
db_dir <- here("data", "raw","SNAC-K")

# this import all csv, sav and dta files in the 'db_dir'
bulk_import(data_dir = db_dir)

# this import only a specific type files
bulk_import(data_dir = db_dir, file_type = 'sav')
}

}
