% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_import_sav.R
\name{bulk_import_csv}
\alias{bulk_import_csv}
\title{Bulk import SPSS(sav) data into R}
\usage{
bulk_import_csv(data_dir = NULL)
}
\arguments{
\item{data_dir}{The folder storing SPSS files.}
}
\value{
1. Import SPSS files in `data_dir` to R global environment;
2. List names of tables imported to R.
}
\description{
Many NEAR databases stroe data in SPSS. Therefore chances are that we need
to import many SPSS files at a time in R for further analysis.
}
\note{
\code{here} function is a fast and neat way of defining data directory `data_dir`.
}
\examples{
\dontrun{
db_dir <- here("data", "raw","SNAC-K")
bulk_import_csv(data_dir = db_dir)
}

}
