% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_dup_id}
\alias{df_dup_id}
\title{Example data frame}
\format{
A data frame with 2 variables: \code{id}, \code{num}.
}
\usage{
df_dup_id
}
\description{
A data frame to test the \code{get_dup_id} function in package.
}
\keyword{datasets}
