% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_caide_df}
\alias{fake_caide_df}
\title{Fake CAIDE data}
\format{
A data frame with 27 variables and 1000 random obs.
}
\usage{
fake_caide_df
}
\description{
A randomized CAIDE data set, with meta data preserved.
}
\details{
Random values are generated by \code{vroom::gen_tbl}.
}
\keyword{datasets}
