% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_snacn_ph_fu}
\alias{fake_snacn_ph_fu}
\title{Fake SNAC-N data 'fake_snacn_ph_fu'}
\format{
A data frame with 4 variables and 800 observations
}
\usage{
fake_snacn_ph_fu
}
\description{
Simulated SNAC-N wave 3 data, but only for cohort 1 follow-up. ID is up to 800.
}
\keyword{datasets}
