% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_snacn_ph_wave3}
\alias{fake_snacn_ph_wave3}
\title{Fake SNAC-N data 'fake_snacn_ph_wave3'}
\format{
A data frame with 4 variables and 800 observations
}
\usage{
fake_snacn_ph_wave3
}
\description{
Simulated SNAC-N wave 3 data, but for the whole wave 3. ID is up to 1000.
}
\keyword{datasets}
