% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_dup_id.R
\name{fix_dup_id}
\alias{fix_dup_id}
\title{Fix replicated id in a data file}
\usage{
fix_dup_id(df)
}
\arguments{
\item{df}{A tibble to be examined.}
}
\value{
A list with examination results
\itemize{
  \item logic_rep: If df has replicated ID or not.
  \item new_lopnr: A new distinct ID. E.g, for both people with ID = 2, the new ID are 2.0, 2.1.
  \item replicated_id: Specific replicated ID. May need to report them to local DBM.
  \item df_update: A new df with new_lopnr appened before original ID column.
}
}
\description{
In NEAR, we use ID (lopnr) to identify a specific participant. If two people share the same ID, there must be a problem.
This function is used to:
* Check if a data file has replicated ID.
* If there is replicated id, find out which IDs are replicated. Generate new distinct new id & append it before the original ID column.
}
\details{
The new ID can be temporarily used as primary key when importing data to MySQL.
}
\examples{
\dontrun{
neartools::fix_dup_id(baseline_example_Relative_220504)$logic_rep
}


}
