% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_csv_na.R
\name{fix_empty_string}
\alias{fix_empty_string}
\title{Fix the empty string in csv file}
\usage{
fix_empty_string(df_name)
}
\arguments{
\item{df_name}{A tibble name in R. It should be a string.}
}
\value{
A tibble with empty strings are replaced as 'NA'.
}
\description{
When importing the csv file to MySQL workbench, if there is a cell with empty strings, the 'Wizard Import' function would skip that row.
To fix this problem, we need to replace the empty string with "NA" so that we can import all records into MySQL.
}
\note{
This replacement only applies to 'character' columns.
}
\examples{
\dontrun{
fix_empty_string("baseline_example_Nurse_220504")
}
}
