% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_label_df.R
\name{get_label_df}
\alias{get_label_df}
\title{Get labels of a dataset}
\usage{
get_label_df(df_w_label)
}
\arguments{
\item{df_w_label}{A data set with labels. One can read SPSS or STATA files with \code{haven} package.}
}
\value{
A tibble with three columns:
  \item{variable}{The original variables' names.}
  \item{label_char}{The variables corresponding labels.}
  \item{na_percent}{Ratio of missing values.}
}
\description{
During work, many times we receive SPSS or STATA data files with labels. This function is to extract the labels from the read-in data, turn them in tibble format.
This can help us to inspect data more conveniently.
}
\examples{
\dontrun{
get_label_df(fake_snack_df)
}

}
