% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_na_number.R
\name{get_na_number}
\alias{get_na_number}
\title{Get NA number of a dataset}
\usage{
get_na_number(
  data_df,
  label_df = NULL,
  keywords_label = NULL,
  keywords_variable = NULL
)
}
\arguments{
\item{data_df}{The data frame, preferably tibble.}

\item{label_df}{The lable of data frame. Get it by \code{neartools::get_label_df(data_df)}.}

\item{keywords_label}{Keywords in the label.}

\item{keywords_variable}{Keywords in the column names.}
}
\value{
A list of related variables and their NA counts.
}
\description{
Given the data set and its label data frame, user can search for number of NA with strings in the variable names or labels.
}
\details{
The \code{grepl()} function is doing the search for the two keywards arguments.
}
\examples{
\dontrun{
label_df <- get_label_df(fake_snack_df)
# search by label
get_na_number(data_df = fake_snack_df, label_df = label_df, keywords_label = "dementia")

# search by variable name
get_na_number(data_df = fake_snack_df, label_df = label_df, keywords_variable = "edu")
}

}
