% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pretty_template.R
\name{get_pretty_template}
\alias{get_pretty_template}
\title{get_pretty_template}
\usage{
get_pretty_template(
  name = NULL,
  output_file = "html",
  open = interactive(),
  ...
)
}
\arguments{
\item{name}{name of rmd file. (no need to add suffix ".Rmd")}

\item{output_file}{"word" or "html", by default it is "html"}

\item{open}{should the file be opened after being created}

\item{...}{arguments to be passed to \link[usethis]{use_template}}
}
\description{
Use the skeleton.Rmd for starting analysis. This function creates a new folder "rmd" at root of project and create an rmd file with nice-looking template there.
}
\examples{
\dontrun{
neartools::get_pretty_template(name = "Reply to Prof")
}

}
