% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_template.R
\name{pretty_template}
\alias{pretty_template}
\title{pretty_template}
\usage{
pretty_template(name = NULL, open = interactive(), ...)
}
\arguments{
\item{name}{name of rmd file. No need to add suffix ".Rmd".}

\item{open}{should the file be opened after being created}

\item{...}{arguments to be passed to \link[usethis]{use_template}}
}
\description{
Use the skeleton.Rmd for starting analysis. This function creates a new folder "rmd" at root of project and create an rmd file with nice-looking template there.
}
\examples{
\donttest{
neartools::pretty_template(name = "Reply to Prof")
}

}
